outsource project() from Approximations
