+++ docs/src/lib/lazy_operations/LinearMap.md
@@ -15,6 +15,7 @@ an_element(::LinearMap)
 vertices_list(::LinearMap)
 constraints_list(::LinearMap)
 linear_map(::AbstractMatrix, ::LinearMap)
+project(::LazySet, ::AbstractVector{Int}, ::Type{<:LinearMap})
 ```
 Inherited from [`AbstractAffineMap`](@ref):
 * [`isempty`](@ref isempty(::AbstractAffineMap))
