+++ docs/src/lib/sets/HalfSpace.md
@@ -27,6 +27,7 @@ tosimplehrep(::AbstractVector{LC}) where {N, LC<:LinearConstraint{N}}
 remove_redundant_constraints
 remove_redundant_constraints!
 complement(::HalfSpace)
+project(::HalfSpace, ::AbstractVector{Int})
 ```
 Inherited from [`LazySet`](@ref):
 * [`norm`](@ref norm(::LazySet, ::Real))
