+++ src/LazyOperations/LinearMap.jl
@@ -460,3 +460,32 @@ function Projection(X::LazySet{N}, variables::AbstractVector{Int}) where {N}
     M = projection_matrix(variables, dim(X), N)
     return LinearMap(M, X)
 end
+
+"""
+    project(S::LazySet{N},
+            block::AbstractVector{Int},
+            set_type::Type{<:LinearMap},
+            [n]::Int=dim(S)
+           ) where {N}
+
+Project a high-dimensional set to a given block by using a lazy linear map.
+
+### Input
+
+- `S`         -- set
+- `block`     -- block structure - a vector with the dimensions of interest
+- `LinearMap` -- used for dispatch
+- `n`         -- (optional, default: `dim(S)`) ambient dimension of the set `S`
+
+### Output
+
+A lazy `LinearMap` representing the projection of the set `S` to block `block`.
+"""
+@inline function project(S::LazySet{N},
+                         block::AbstractVector{Int},
+                         set_type::Type{<:LinearMap},
+                         n::Int=dim(S)
+                        ) where {N}
+    M = projection_matrix(block, n, N)
+    return M * S
+end
