+++ src/Sets/HalfSpace.jl
@@ -672,3 +672,85 @@ of ``H`` and ``H′`` is non-empty).
 function complement(H::HalfSpace)
     return HalfSpace(-H.a, -H.b)
 end
+
+"""
+    project(H::HalfSpace, block::AbstractVector{Int})
+
+Concrete projection of a half-space.
+
+### Input
+
+- `H`        -- set
+- `block`    -- block structure, a vector with the dimensions of interest
+
+### Output
+
+A set representing the projection of the half-space `H` on the dimensions
+specified by `block`.
+
+### Algorithm
+
+If the unconstrained dimensions of `H` are a subset of the `block` variables,
+the projection is applied to the normal direction of `H`.
+Otherwise, the projection results in the universal set.
+
+The latter can be seen as follows.
+Without loss of generality consider a projection onto a single and constrained
+dimension ``xₖ`` (projections in multiple dimensions can be modeled as repeated
+one-dimensional projections).
+We can write the projection as an existentially quantified linear constraint:
+
+```math
+    ∃xₖ: a₁x₁ + … + aₖxₖ + … + aₙxₙ ≤ b
+```
+
+Since ``aₖ ≠ 0``, there is always a value for ``xₖ`` that satisfies the
+constraint for any valuation of the other variables.
+
+### Examples
+
+Consider the half-space ``x + y + 0⋅z ≤ 1``, whose ambient dimension is `3`.
+The (trivial) projection in the three dimensions is achieved letting the block
+of variables to be `[1, 2, 3]`:
+
+```jldoctest project_halfspace
+julia> H = HalfSpace([1.0, 1.0, 0.0], 1.0)
+HalfSpace{Float64,Array{Float64,1}}([1.0, 1.0, 0.0], 1.0)
+
+julia> project(H, [1, 2, 3])
+HalfSpace{Float64,Array{Float64,1}}([1.0, 1.0, 0.0], 1.0)
+```
+
+Projecting along dimensions `1` and `2` only:
+
+```jldoctest project_halfspace
+julia> project(H, [1, 2])
+HalfSpace{Float64,Array{Float64,1}}([1.0, 1.0], 1.0)
+```
+
+In general, use the call syntax `project(H, constrained_dimensions(H))` to return
+the half-space projected on the dimensions where it is constrained only:
+
+```jldoctest project_halfspace
+julia> project(H, constrained_dimensions(H))
+HalfSpace{Float64,Array{Float64,1}}([1.0, 1.0], 1.0)
+```
+
+If a constrained dimension is projected, we get the universal set of the
+dimension corresponding to the projection.
+
+```jldoctest project_halfspace
+julia> project(H, [1, 3])
+Universe{Float64}(2)
+
+julia> project(H, [1])
+Universe{Float64}(1)
+```
+"""
+function project(H::HalfSpace, block::AbstractVector{Int})
+    if constrained_dimensions(H) ⊆ block
+        return HalfSpace(H.a[block], H.b)
+    else
+        return Universe(length(block))
+    end
+end
