+++ src/Sets/VPolytope.jl
@@ -544,3 +544,23 @@ end
 
 end # quote
 end # function load_polyhedra_vpolytope()
+
+function project(V::Union{<:VPolygon{N}, <:VPolytope{N}},
+                 block::AbstractVector{Int}) where {N}
+    n = dim(V)
+    M = projection_matrix(block, n, N)
+    πvertices = broadcast(v -> M * v, vertices_list(V))
+
+    m = size(M, 1)
+    if m == 1
+        # convex_hull in 1d returns the minimum and maximum points, in that order
+        aux = convex_hull(πvertices)
+        a = first(aux[1])
+        b = length(aux) == 1 ? a : first(aux[2])
+        return Interval(a, b)
+    elseif m == 2
+        return VPolygon(πvertices; apply_convex_hull=true)
+    else
+        return VPolytope(πvertices)
+    end
+end
